+++ R/plans.R
@@ -203,7 +203,7 @@ get_plan <- function(countries, additional_exemplar_countries = NULL,
                                                                      completed_fu_allocation_tables = CompletedAllocationTables,
                                                                      countries = countries,
                                                                      max_year = max_year),
-                                              dynamic = map(countries))
+                                              dynamic = map(countries)),
 
     # (10) Extend to useful stage
 
@@ -223,7 +223,16 @@ get_plan <- function(countries, additional_exemplar_countries = NULL,
 
 
     # (N) Build reports
+
+    # Build Allocation Graphs
+    AllocationGraphs = drake::target(alloc_plots_df(CompletedAllocationTables, countries = countries),
+                                     dynamic = map(countries)),
     # Allocation_Report =
+
+    # Build Efficiency Graphs
+    EfficiencyGraphs = drake::target(eta_fu_plots_df(CompletedEfficiencyTables, countries = countries))
+                       #,dynamic = map(machine & eu_product)) # How to map by
+
     # reports_source_paths = drake::target(drake::file_in(report_source_paths(report_source_folders = report_source_folders))),
     # reports_dest_path = drake::target(drake::file_out(report_dest_paths(report_source_paths))),
     # reports_complete = drake::target(generate_reports(report_source_files = report_source_paths,