+++ NEWS.md
@@ -1,3 +1,24 @@
+# SEAPSUTWorkflow 0.1.9 (2020-09-17)
+
+* New functions `eta_fu_graph()` and `eta_fu_plots_df()`
+  which make efficiency graphs for later use in reports and other
+  visualization.
+
+
+# SEAPSUTWorkflow 0.1.8 (2020-09-09)
+
+* Changed structure of data frame in AllocationGraphs target. 
+  Now, fewer columns are in the data that is graphed. 
+  More columns are in the outer (not nested) data frame.
+
+
+# SEAPSUTWorkflow 0.1.7 (2020-09-07)
+
+* Added allocation report.
+* New functions `alloc_graph()` and `alloc_plots_df()`
+  which make allocation graphs for later use in reports.
+
+
 # SEAPSUTWorkflow 0.1.6 (2020-08-26)
 
 * First version that works with full SEAPSUT workflow.